+++ test/libsolidity/smtCheckerTests/types/array_literal_3.sol
@@ -11,5 +11,5 @@ contract C
 // SMTEngine: all
 // SMTIgnoreCex: no
 // ----
-// Warning 6328: (168-188): CHC: Assertion violation happens here.
+// Warning 6328: (168-188): CHC: Assertion violation happens here.\nCounterexample:\n\na = [1, 2, 3]\nb = [1, 2, 4]\n\nTransaction trace:\nC.constructor()\nC.f()
 // Info 1391: CHC: 8 verification condition(s) proved safe! Enable the model checker option "show proved safe" to see all of them.